+++ Sources/SmileID/Classes/SmileID.swift
@@ -87,7 +87,28 @@ public class SmileID {
         jobId: String,
         deleteFilesOnSuccess: Bool = false
     ) {
-        OfflineMode.submitJob(jobId: jobId, deleteFilesOnSuccess: deleteFilesOnSuccess)
+        static func submitJob(
+            jobId: String,
+            deleteFilesOnSuccess: Bool // todo - to be used in next pr
+        ) {
+            Task {
+                do {
+                    let authRequest = try LocalStorage.fetchAuthenticationRequestFile(jobId: jobId)
+                    let authResponse = try await SmileID.api.authenticate(
+                        request: authRequest
+                    ).async()
+                    let prepUploadRequest = try LocalStorage.fetchPrepUploadFile(jobId: jobId)
+                    let prepUploadResponse = try await SmileID.api.prepUpload(
+                        request: prepUploadRequest
+                    ).async()
+                    let zip = try LocalStorage.fetchUploadZip(jobId: jobId)
+                    let response = try await SmileID.api.upload(
+                        zip: zip,
+                        to: prepUploadResponse.uploadUrl
+                    ).async()
+                }
+            }
+        }
     }
 
     /// deletes the job ids list, whether in unsubmitted or submitted state
